Fixed styles on events
