--- src/components/HeaderWithSearchBar/index.js
@@ -1,3 +1,0 @@
-import HeaderWithSearchBar from './HeaderWithSearchBar.vue';
-
-export default HeaderWithSearchBar;
