+++ src/components/ListHeader/index.js
@@ -1,0 +1,3 @@
+import ListHeader from './ListHeader.vue';
+
+export default ListHeader;
