+++ src/pages/Events/Events.layout.vue
@@ -1,5 +1,5 @@
 <template>
-  <header-with-search-bar
+  <list-header
     title="Eventos"
     :query-value="searchText || ''"
     @update:query-value="$emit('update:searchText', $event?.toString())"
@@ -28,9 +28,9 @@
 
 <script lang="ts" setup>
 import { Event, EventType } from 'src/utils';
-import { HeaderWithSearchBar } from 'src/components';
+import { ListHeader } from 'src/components';
 import EventCard from 'src/components/EventCard';
-import DIDLogo from 'src/assets/logos/didperu.svg';
+import DIDLogo from 'src/assets/logos/didperu-dark.svg';
 import './styles.scss';
 import BaseLoading from 'src/components/LoadingComponent/BaseLoading.vue';
