+++ src/pages/Events/styles.scss
@@ -2,11 +2,11 @@
   &__list {
     display: flex;
     flex-direction: column;
-    gap: 0.5rem;
+    gap: 0.625rem;
   }
   &__container {
-    padding: 0 2.31rem 1.12rem;
+    padding: 0 3.375rem 1.12rem 2.25rem;
     margin: auto;
-    margin-top: 0.875rem;
+    margin-top: 2.2rem;
   }
 }
